2-data-simulator/requirements.txt
pandas==2.0.3
numpy==1.24.3
psycopg2-binary==2.9.9